+++ apps/api/package.json
@@ -18,7 +18,8 @@
     "@types/express": "^5.0.0",
     "@types/node": "^22.13.1",
     "tsc-watch": "^6.2.1",
-    "typescript": "^5.7.3"
+    "typescript": "^5.7.3",
+    "eslint": "^8.57.1"
   },
   "dependencies": {
     "better-auth": "^1.1.17",
