+++ packages/db/package.json
@@ -16,7 +16,8 @@
     "ts-node": "^10.9.2",
     "typescript": "^5.7.3",
     "@repo/typescript-config": "workspace:*",
-    "@repo/eslint-config": "workspace:*"
+    "@repo/eslint-config": "workspace:*",
+    "eslint": "^8.57.1"
   },
   "dependencies": {
     "@prisma/client": "^6.3.1"
