+++ packages/ui/package.json
@@ -16,7 +16,8 @@
     "react": "^19.0.0",
     "tailwind-merge": "^3.0.1",
     "tailwindcss": "^4.0.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "eslint": "^8.57.1"
   },
   "dependencies": {
     "@radix-ui/react-alert-dialog": "^1.1.6",
